refactor: move metrics to @streamr/utils
