+++ packages/broker/src/plugins/consoleMetrics/ConsoleMetricsPlugin.ts
@@ -1,8 +1,7 @@
 import { Schema } from 'ajv'
 import { Plugin } from '../../Plugin'
 import PLUGIN_CONFIG_SCHEMA from './config.schema.json'
-import { MetricsReport } from 'streamr-network'
-import { Logger } from '@streamr/utils'
+import { Logger, MetricsReport } from '@streamr/utils'
 import { omit } from 'lodash'
 
 const logger = new Logger(module)
