+++ packages/broker/src/plugins/storage/DataQueryEndpoints.ts
@@ -3,8 +3,7 @@
  */
 import express, { Request, Response, Router } from 'express'
 import { StreamMessage } from 'streamr-client-protocol'
-import { Metric, MetricsContext, RateMetric } from 'streamr-network'
-import { Logger } from '@streamr/utils'
+import { Logger, Metric, MetricsContext, RateMetric } from '@streamr/utils'
 import { Readable, Transform, pipeline } from 'stream'
 import { Storage } from './Storage'
 import { Format, getFormat } from './DataQueryFormat'
