+++ packages/network/src/browser.ts
@@ -1,9 +1,6 @@
 import 'setimmediate'
 export { NameDirectory } from './NameDirectory'
-export { Logger } from '@streamr/utils'
-export { 
-    MetricsContext
-} from './helpers/Metric'
+export { Logger, MetricsContext } from '@streamr/utils'
 export { Location, AbstractNodeOptions } from './identifiers'
 export { createNetworkNode, NetworkNodeOptions } from './createNetworkNode'
 export { NetworkNode } from './logic/NetworkNode'
