+++ packages/network/test/integration/delivery-of-messages-protocol-layer.test.ts
@@ -20,7 +20,7 @@ import { Event as NodeToTrackerEvent, NodeToTracker } from '../../src/protocol/N
 import { PeerInfo } from '../../src/connection/PeerInfo'
 import { RtcSignaller } from '../../src/logic/RtcSignaller'
 import { NegotiatedProtocolVersions } from '../../src/connection/NegotiatedProtocolVersions'
-import { MetricsContext } from '../../src/helpers/Metric'
+import { MetricsContext } from '@streamr/utils'
 import { WebRtcEndpoint } from '../../src/connection/webrtc/WebRtcEndpoint'
 import { webRtcConnectionFactory } from '../../src/connection/webrtc/NodeWebRtcConnection'
 import NodeClientWsEndpoint from '../../src/connection/ws/NodeClientWsEndpoint'
